Make command chaining safer.
When we need to send commands that must strictly follow each-other, the promises are a bit broken since they cannot guarantee nobody will come in-between. So, add a helper for the particular case where we want a command to be run in the context of a group. Also, ensure we have the ability to post several commands at once.